mvp_requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
pillow==10.2.0
python-docx==1.1.0
requests==2.31.0
google-generativeai==0.3.2
numpy==1.26.3
